Add error logging and update QnaItem id type
Added error logging to chatingRoomVM for better debugging of message sending failures. Changed QnaItem id type from Int to Long in qnaVM to support larger identifier values.

+++ app/src/main/java/com/example/studify/ui/qnaVM.kt
@@ -83,5 +83,5 @@ class qnaVM @Inject constructor(
         disposables.clear()
     }
 
-    data class QnaItem(val id: Int, val title: String, val content: String, val answer: String)
+    data class QnaItem(val id: Long, val title: String, val content: String, val answer: String)
 }
